Add basic smoke test with fake client
