+++ androidApp/src/test/kotlin/io/music_assistant/client/feature/SmokeTest.kt
@@ -1,0 +1,24 @@
+package io.music_assistant.client.feature
+
+import androidx.compose.ui.test.ExperimentalTestApi
+import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.music_assistant.client.ui.compose.App
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class SmokeTest {
+
+    @get:Rule
+    val composeTestRule = createComposeRule()
+
+    @OptIn(ExperimentalTestApi::class)
+    @Test
+    fun `can launch app`() {
+        composeTestRule.setContent {
+            App()
+        }
+    }
+}
