+++ androidApp/src/test/resources/robolectric.properties
@@ -1,1 +1,2 @@
 sdk=34
+application=io.music_assistant.client.support.TestApplication
