+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/di/SharedModule.kt
@@ -1,5 +1,6 @@
 package io.music_assistant.client.di
 
+import io.music_assistant.client.api.KtorServiceClient
 import io.music_assistant.client.api.ServiceClient
 import io.music_assistant.client.auth.AuthenticationManager
 import io.music_assistant.client.data.LocalPlayerRepository
@@ -18,35 +19,37 @@ import io.music_assistant.client.ui.compose.search.SearchViewModel
 import io.music_assistant.client.ui.compose.settings.SettingsViewModel
 import io.music_assistant.client.ui.theme.ThemeViewModel
 import io.music_assistant.client.utils.NetworkMonitor
+import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
-val sharedModule = module {
-    single { provideSettings() }
-    singleOf(::SettingsRepository)
-    singleOf(::NetworkMonitor)
-    singleOf(::ServiceClient)
-    singleOf(::LogSharer)
-    single(createdAtStart = true) {
-        AuthenticationManager(
-            get(),
-            get()
-        )
-    }  // Eager - needs to start monitoring immediately
-    singleOf(::MediaPlayerController)  // Used by MainDataSource for Sendspin
-    singleOf(::SendspinClientFactory)   // Factory for creating Sendspin clients
-    singleOf(::LocalPlayerRepository)   // Optimistic local player state
-    singleOf(::MainDataSource)          // Singleton - held by foreground service
-    viewModelOf(::ThemeViewModel)
-    factory { ActionsViewModel(get(), get()) }
-    factory { SettingsViewModel(get(), get(), get()) }
-    factory { AuthenticationViewModel(get(), get()) }
-    factory { LibraryViewModel(get(), get(), get()) }
-    factory { ItemDetailsViewModel(get(), get(), get()) }
-    factory { HomeScreenViewModel(get(), get(), get()) }
-    factory { SearchViewModel(get(), get()) }
-}
+fun sharedModule(serviceClientConstructor: (SettingsRepository) -> ServiceClient = ::KtorServiceClient) =
+    module {
+        single { provideSettings() }
+        singleOf(::SettingsRepository)
+        singleOf(::NetworkMonitor)
+        singleOf(serviceClientConstructor) { bind<ServiceClient>() }
+        singleOf(::LogSharer)
+        single(createdAtStart = true) {
+            AuthenticationManager(
+                get(),
+                get()
+            )
+        }  // Eager - needs to start monitoring immediately
+        singleOf(::MediaPlayerController)  // Used by MainDataSource for Sendspin
+        singleOf(::SendspinClientFactory)   // Factory for creating Sendspin clients
+        singleOf(::LocalPlayerRepository)   // Optimistic local player state
+        singleOf(::MainDataSource)          // Singleton - held by foreground service
+        viewModelOf(::ThemeViewModel)
+        factory { ActionsViewModel(get(), get()) }
+        factory { SettingsViewModel(get(), get(), get()) }
+        factory { AuthenticationViewModel(get(), get()) }
+        factory { LibraryViewModel(get(), get(), get()) }
+        factory { ItemDetailsViewModel(get(), get(), get()) }
+        factory { HomeScreenViewModel(get(), get(), get()) }
+        factory { SearchViewModel(get(), get()) }
+    }
 
 /**
  * Cleanup function to properly close all singleton resources.
