+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/di/initKoin.kt
@@ -1,6 +1,7 @@
 package io.music_assistant.client.di
 
 import co.touchlab.kermit.Logger
+import io.music_assistant.client.api.KtorServiceClient
 import io.music_assistant.client.logging.InMemoryLogWriter
 import org.koin.core.context.startKoin
 import org.koin.core.module.Module
@@ -13,6 +14,6 @@ fun initKoin(
     Logger.addLogWriter(InMemoryLogWriter)
     startKoin {
         config?.invoke(this)
-        modules(sharedModule, webrtcModule, *platformModules)
+        modules(sharedModule(), webrtcModule, *platformModules)
     }
 }
